+++ index.js
@@ -1,12 +1,16 @@
 const {
   ProgressBar,
+  Spinner,
   presets,
+  spinnerStyles,
   effects
 } = require('./lib');
 
 module.exports = ProgressBar;
 
 module.exports.ProgressBar = ProgressBar;
+module.exports.Spinner = Spinner;
 module.exports.presets = presets;
+module.exports.spinnerStyles = spinnerStyles;
 module.exports.effects = effects;
 module.exports.default = ProgressBar;